+++ lib/routes/transaction_page.dart
@@ -21,7 +21,6 @@ import "package:flow/routes/new_transaction/select_category_sheet.dart";
 import "package:flow/routes/new_transaction/title_input.dart";
 import "package:flow/services/exchange_rates.dart";
 import "package:flow/services/transactions.dart";
-import "package:flow/services/user_preferences.dart";
 import "package:flow/theme/theme.dart";
 import "package:flow/utils/utils.dart";
 import "package:flow/widgets/general/button.dart";
@@ -600,9 +599,7 @@ class _TransactionPageState extends State<TransactionPage> {
 
     if (!mounted) return;
 
-    if (widget.isNewTransaction &&
-        result != null &&
-        !UserPreferencesService().ignoreTitlesInRecordFlow) {
+    if (widget.isNewTransaction && result != null) {
       FocusScope.of(context).requestFocus(_titleFocusNode);
     }
   }